refactor: move public types to /@api/ Signed-off-by: Philippe Martin <phmartin@redhat.com>
