+++ packages/main/src/plugin/kubernetes/kubernetes-client.ts
@@ -68,6 +68,7 @@ import * as jsYaml from 'js-yaml';
 import { parseAllDocuments } from 'yaml';
 
 import type { KubeContext } from '/@api/kubernetes-context.js';
+import type { ContextGeneralState, ResourceName } from '/@api/kubernetes-contexts-states.js';
 import type { V1Route } from '/@api/openshift-types.js';
 
 import type { ApiSenderType } from '../api.js';
@@ -77,7 +78,6 @@ import { Emitter } from '../events/emitter.js';
 import type { FilesystemMonitoring } from '../filesystem-monitoring.js';
 import type { Telemetry } from '../telemetry/telemetry.js';
 import { Uri } from '../types/uri.js';
-import type { ContextGeneralState, ResourceName } from './contexts-states.js';
 import { ContextsManager } from './kubernetes-context-state.js';
 import { BufferedStreamWriter, ResizableTerminalWriter, StringLineReader } from './kubernetes-exec-transmitter.js';
