+++ packages/renderer/src/lib/preferences/PreferencesKubernetesContextsRendering.spec.ts
@@ -25,8 +25,8 @@ import { beforeEach, expect, test, vi } from 'vitest';
 import { kubernetesContexts } from '/@/stores/kubernetes-contexts';
 import * as kubernetesContextsState from '/@/stores/kubernetes-contexts-state';
 import type { KubeContext } from '/@api/kubernetes-context';
+import type { CheckingState, ContextGeneralState } from '/@api/kubernetes-contexts-states';
 
-import type { CheckingState, ContextGeneralState } from '../../../../main/src/plugin/kubernetes/contexts-states';
 import PreferencesKubernetesContextsRendering from './PreferencesKubernetesContextsRendering.svelte';
 
 vi.mock('/@/stores/kubernetes-contexts-state', async () => {
