+++ packages/renderer/src/lib/ui/KubernetesCurrentContextConnectionBadge.spec.ts
@@ -24,8 +24,7 @@ import { render, screen } from '@testing-library/svelte';
 import { beforeEach, expect, test, vi } from 'vitest';
 
 import KubernetesCurrentContextConnectionBadge from '/@/lib/ui/KubernetesCurrentContextConnectionBadge.svelte';
-
-import type { ContextGeneralState } from '../../../../main/src/plugin/kubernetes/contexts-states';
+import type { ContextGeneralState } from '/@api/kubernetes-contexts-states';
 
 const mocks = vi.hoisted(() => ({
   subscribeMock: vi.fn(),
